+++ src/domain/packages/validator/validations/o-mq-2/__tests__/data/O_MQ_2_success.json
