+++ src/domain/packages/validator/validations/o-mq-2/__tests__/validate_O_MQ_2.spec.ts
