+++ src/domain/packages/validator/validations/o-mq-2/validate_O_MQ_2.ts
