+++ src/domain/packages/validator/validations/og-mq-1/__tests__/data/OG_MQ_1_success.json
