+++ src/domain/packages/validator/validations/og-mq-1/validate-OG-MQ-1.ts
