+++ src/domain/packages/validator/validations/pr-st-3/__tests__/data/PR_ST_3_success.json
