+++ src/domain/packages/validator/validations/pr-st-3/__tests__/validate_PR_ST_3.spec.ts
