+++ src/domain/packages/validator/validations/pr-st-3/validate_PR_ST_3.ts
