+++ src/domain/packages/validator/validations/shst-mq-1/__tests__/data/SHST_MQ_1_fail.json
