+++ src/domain/packages/validator/validations/shst-mq-1/__tests__/validate_SHST_MQ_1.spec.ts
