+++ src/domain/packages/validator/validations/shst-mq-1/validate_SHST_MQ_1.ts
